Extracted Contact class with Address, Email and Phone as subclasses
